Change `tmux-start` to be more comprehensive
tmux-start now creates two panes and starts vi in one, git in the other.
It also gained some intelligence on whether the session already exists
(at which point it will just attach to it) and whether it is being run
from within Tmux or not (which affects whether we attach to the session
or switch to it).